+++ xxyp/src/main/java/com/xxyp/model/WorksPhotoExample.java
@@ -484,7 +484,7 @@ public class WorksPhotoExample {
             return (Criteria) this;
         }
 
-        public Criteria andStatusEqualTo(Byte value) {
+        public Criteria andStatusEqualTo(Integer value) {
             addCriterion("status =", value, "status");
             return (Criteria) this;
         }
